+++ src/overlay/StellarXDR.h
@@ -1,9 +1,9 @@
 #pragma once
-#include "xdr/Stellar-ledger-entries.h"
-#include "xdr/Stellar-ledger.h"
-#include "xdr/Stellar-overlay.h"
-#include "xdr/Stellar-transaction.h"
-#include "xdr/Stellar-types.h"
+#include "xdr/ledger-entries.h"
+#include "xdr/ledger.h"
+#include "xdr/overlay.h"
+#include "xdr/transaction.h"
+#include "xdr/types.h"
 
 namespace stellar
 {