+++ src/transactions/test/test_helper/ManageAMLAlertTestHelper.h
@@ -4,7 +4,7 @@
 #include "TxHelper.h"
 #include <ledger/AssetPairFrame.h>
 #include <ledger/AssetPairHelper.h>
-#include <transactions/ManageAssetPairOpFrame.h>
+#include <transactions/manage_asset_pair/ManageAssetPairOpFrame.h>
 
 namespace stellar
 {